+++ packages/legend-application-query-bootstrap/src/index.tsx
@@ -21,8 +21,9 @@ import {
   WebConsole,
 } from '@finos/legend-shared';
 import {
-  DSL_DataSpace_GraphManagerPreset as DSL_DataSpace_GraphManagerPreset,
-  DSL_DataSpace_LegendQueryApplicationPlugin as DSL_DataSpace_LegendQueryApplicationPlugin,
+  DSL_DataSpace_GraphManagerPreset,
+  DSL_DataSpace_LegendQueryApplicationPlugin,
+  DSL_DataSpace_LegendApplicationPlugin,
 } from '@finos/legend-extension-dsl-data-space';
 import { DSL_Text_GraphManagerPreset } from '@finos/legend-extension-dsl-text';
 import { DSL_Diagram_GraphManagerPreset } from '@finos/legend-extension-dsl-diagram';
@@ -33,6 +34,7 @@ import { FMT_JSONSchema_GraphManagerPreset } from '@finos/legend-extension-forma
 import { DSL_Persistence_GraphManagerPreset } from '@finos/legend-extension-dsl-persistence';
 import { DSL_Mastery_GraphManagerPreset } from '@finos/legend-extension-dsl-mastery';
 import { DSL_Service_LegendQueryApplicationPlugin } from '@finos/legend-extension-dsl-service';
+import { FMT_GraphQL_GraphManagerPreset } from '@finos/legend-extension-format-graphql';
 
 export class LegendQueryWebApplication {
   static getPresetCollection(): AbstractPreset[] {
@@ -47,6 +49,7 @@ export class LegendQueryWebApplication {
       new DSL_PersistenceCloud_GraphManagerPreset(),
       new FMT_JSONSchema_GraphManagerPreset(),
       new STO_ServiceStore_GraphManagerPreset(),
+      new FMT_GraphQL_GraphManagerPreset(),
     ];
   }
 
@@ -58,6 +61,8 @@ export class LegendQueryWebApplication {
 
       // loggers
       new WebConsole(),
+      // generic dataspace plugin
+      new DSL_DataSpace_LegendApplicationPlugin(),
     ];
   }