+++ packages/legend-application-query/src/stores/QueryEditorStore.ts
@@ -44,7 +44,6 @@ import {
   PackageableElementExplicitReference,
   RuntimePointer,
   GRAPH_MANAGER_EVENT,
-  GraphManagerTelemetry,
   extractElementNameFromPath,
   QuerySearchSpecification,
   Mapping,
@@ -76,6 +75,7 @@ import {
   TAB_SIZE,
   DEFAULT_TYPEAHEAD_SEARCH_MINIMUM_SEARCH_LENGTH,
   DEFAULT_TYPEAHEAD_SEARCH_LIMIT,
+  ApplicationTelemetry,
 } from '@finos/legend-application';
 import type { LegendQueryPluginManager } from '../application/LegendQueryPluginManager.js';
 import { LegendQueryEventService } from './LegendQueryEventService.js';
@@ -534,26 +534,20 @@ export abstract class QueryEditorStore {
     // report
     stopWatch.record(GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH__SUCCESS);
     const graphBuilderReportData = {
-      timings: {
-        [GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH_SYSTEM__SUCCESS]:
-          stopWatch.getRecord(
-            GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH_SYSTEM__SUCCESS,
-          ),
-        [GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH__SUCCESS]: stopWatch.getRecord(
-          GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH__SUCCESS,
-        ),
-      },
+      timings:
+        this.applicationStore.timeService.finalizeTimingsRecord(stopWatch),
       dependencies: dependency_buildReport,
       dependenciesCount:
         this.graphManagerState.graph.dependencyManager.numberOfDependencies,
       graph: graph_buildReport,
     };
-    this.applicationStore.log.info(
-      LogEvent.create(GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH__SUCCESS),
+    ApplicationTelemetry.logEvent_GraphInitializationSucceeded(
+      this.applicationStore.telemetryService,
       graphBuilderReportData,
     );
-    GraphManagerTelemetry.logEvent_GraphInitializationSucceeded(
-      this.applicationStore.telemetryService,
+
+    this.applicationStore.log.info(
+      LogEvent.create(GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH__SUCCESS),
       graphBuilderReportData,
     );
   }