+++ packages/legend-application-studio/src/stores/EditorGraphState.ts
@@ -90,7 +90,6 @@ import {
   DependencyGraphBuilderError,
   GraphDataDeserializationError,
   GraphBuilderError,
-  GraphManagerTelemetry,
   DataElement,
   type PackageableElement,
   type CompilationWarning,
@@ -103,6 +102,7 @@ import {
 import {
   ActionAlertActionType,
   ActionAlertType,
+  ApplicationTelemetry,
   type TabState,
 } from '@finos/legend-application';
 import { CONFIGURATION_EDITOR_TAB } from './editor-state/project-configuration-editor-state/ProjectConfigurationEditorState.js';
@@ -403,11 +403,10 @@ export class EditorGraphState {
       // report
       stopWatch.record(GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH__SUCCESS);
       const graphBuilderReportData = {
-        timings: {
-          [GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH__SUCCESS]: stopWatch.getRecord(
-            GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH__SUCCESS,
+        timings:
+          this.editorStore.applicationStore.timeService.finalizeTimingsRecord(
+            stopWatch,
           ),
-        },
         dependencies: dependency_buildReport,
         dependenciesCount:
           this.editorStore.graphManagerState.graph.dependencyManager
@@ -416,12 +415,13 @@ export class EditorGraphState {
         generations: generation_buildReport,
         generationsCount: this.graphGenerationState.generatedEntities.size,
       };
-      this.editorStore.applicationStore.log.info(
-        LogEvent.create(GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH__SUCCESS),
+      ApplicationTelemetry.logEvent_GraphInitializationSucceeded(
+        this.editorStore.applicationStore.telemetryService,
         graphBuilderReportData,
       );
-      GraphManagerTelemetry.logEvent_GraphInitializationSucceeded(
-        this.editorStore.applicationStore.telemetryService,
+
+      this.editorStore.applicationStore.log.info(
+        LogEvent.create(GRAPH_MANAGER_EVENT.INITIALIZE_GRAPH__SUCCESS),
         graphBuilderReportData,
       );
 
@@ -636,7 +636,11 @@ export class EditorGraphState {
         }
       }
 
-      report.timings.total = stopWatch.elapsed;
+      report.timings =
+        this.editorStore.applicationStore.timeService.finalizeTimingsRecord(
+          stopWatch,
+          report.timings,
+        );
       LegendStudioTelemetry.logEvent_GraphCompilationSucceeded(
         this.editorStore.applicationStore.telemetryService,
         report,
@@ -809,12 +813,12 @@ export class EditorGraphState {
         ),
       );
 
-      report.timings = {
-        ...report.timings,
-        ...Object.fromEntries(stopWatch.records),
-        total: stopWatch.elapsed,
-      };
-      LegendStudioTelemetry.logEvent_GraphCompilationSucceeded(
+      report.timings =
+        this.editorStore.applicationStore.timeService.finalizeTimingsRecord(
+          stopWatch,
+          report.timings,
+        );
+      LegendStudioTelemetry.logEvent_TextCompilationSucceeded(
         this.editorStore.applicationStore.telemetryService,
         report,
       );