+++ packages/legend-application/src/components/LegendApplicationComponentFrameworkProvider.tsx
@@ -28,6 +28,7 @@ import {
   type KeyBindingConfig,
 } from '@finos/legend-shared';
 import { VirtualAssistant } from './VirtualAssistant.js';
+import { ApplicationTelemetry } from '../stores/ApplicationTelemetry.js';
 
 const APP_CONTAINER_ID = 'app.container';
 const APP_BACKDROP_CONTAINER_ID = 'app.backdrop-container';
@@ -37,17 +38,21 @@ const PLATFORM_NATIVE_KEYBOARD_SHORTCUTS = [
   'Control+KeyP',
   'Meta+KeyS', // Save
   'Control+KeyS',
+
   'F8', // Chrome: Developer Tools > Sources: Run or pause script
   'F10', // Chrome: Developer Tools > Sources: Step over next function call
   'F11', // Chrome: Developer Tools > Sources: Step into next function call
   'Meta+Shift+KeyP', // Chrome: Developer Tools: Open Command Prompt inside developer tools
   'Control+Backquote', // Chrome: Developer Tools: Open console
-  'Control+Shift+KeyP',
+  'Control+Shift+KeyN', // Chrome: Open Private Browsing (incognito)
+
+  'Control+Shift+KeyP', // Firefox: Open Private Browsing
   'Meta+KeyB', // Firefox: Open bookmark sidebar
   'Control+KeyB',
   'F7', // Firefox: Caret browsing
   'Alt+F7', // Firefox: Caret browsing (Mac)
   'Control+Shift+KeyB',
+
   'Alt+KeyZ', // Mac: special symbol
 ];
 
@@ -66,6 +71,8 @@ const buildHotkeysConfiguration = (
   });
 
   // Disable platform native keyboard shortcuts
+  // NOTE: due to the order in which hotkey configuration is searched and applied,
+  // we must place these after application hotkey configuration
   const PLATFORM_NATIVE_KEYBOARD_COMMAND =
     'INTERNAL__PLATFORM_NATIVE_KEYBOARD_COMMAND';
   keyMap[PLATFORM_NATIVE_KEYBOARD_COMMAND] = {
@@ -136,6 +143,29 @@ export const LegendApplicationComponentFrameworkProvider = observer(
         document.removeEventListener('keydown', onKeyEvent);
       };
     }, [keyBindingMap]);
+
+    /**
+     * Keep track of when the application usage is interrupted (e.g. when the app window/tab is not in focus),
+     * since for certain platform, background contexts are de-prioritized, given less resources, and hence, would
+     * run less performantly; and might require particular handlings.
+     *
+     * See https://developer.mozilla.org/en-US/docs/Web/API/Page_Visibility_API#policies_in_place_to_aid_background_page_performance
+     * See https://plumbr.io/blog/performance-blog/background-tabs-in-browser-load-20-times-slower
+     */
+    useEffect(() => {
+      const onVisibilityChange = (): void => {
+        if (document.hidden) {
+          ApplicationTelemetry.logEvent_ApplicationUsageInterrupted(
+            applicationStore.telemetryService,
+          );
+          applicationStore.timeService.recordUsageInterruption();
+        }
+      };
+      document.addEventListener('visibilitychange', onVisibilityChange);
+      return () => {
+        document.removeEventListener('visibilitychange', onVisibilityChange);
+      };
+    }, [applicationStore]);
 
     return (
       <LegendStyleProvider>