+++ packages/legend-application/src/index.ts
@@ -36,6 +36,8 @@ export * from './stores/storage/StorageService.js';
 export * from './stores/DocumentationService.js';
 export * from './stores/CommandCenter.js';
 export * from './stores/EventService.js';
+export * from './stores/TelemetryService.js';
+export * from './stores/IframeEventNotifierPlugin.js';
 export * from './stores/NotificationService.js';
 export * from './stores/AlertService.js';
 export * from './stores/AssistantService.js';