+++ packages/legend-application/src/stores/IframeEventNotifierPlugin.ts
@@ -18,7 +18,7 @@ import packageJson from '../../package.json';
 import {
   type NotificationEventData,
   EventNotifierPlugin,
-} from './EventNotifierPlugin.js';
+} from './EventService.js';
 
 export class IframeEventNotifierPlugin extends EventNotifierPlugin {
   constructor() {