+++ packages/legend-query-builder/src/stores/QueryBuilderResultState.ts
@@ -256,10 +256,11 @@ export class QueryBuilderResultState {
         this.latestRunHashCode = currentHashCode;
         this.setExecutionDuration(stopWatch.elapsed);
 
-        report.timings = {
-          ...report.timings,
-          total: stopWatch.elapsed,
-        };
+        report.timings =
+          this.queryBuilderState.applicationStore.timeService.finalizeTimingsRecord(
+            stopWatch,
+            report.timings,
+          );
         QueryBuilderTelemetry.logEvent_QueryRunSucceeded(
           this.queryBuilderState.applicationStore.telemetryService,
           report,
@@ -349,11 +350,12 @@ export class QueryBuilderResultState {
       }
       stopWatch.record(QUERY_BUILDER_EVENT.BUILD_EXECUTION_PLAN__SUCCESS);
 
-      report.timings = {
-        ...report.timings,
-        ...Object.fromEntries(stopWatch.records),
-        total: stopWatch.elapsed,
-      };
+      // report
+      report.timings =
+        this.queryBuilderState.applicationStore.timeService.finalizeTimingsRecord(
+          stopWatch,
+          report.timings,
+        );
       if (debug) {
         QueryBuilderTelemetry.logEvent_ExecutionPlanDebugSucceeded(
           this.queryBuilderState.applicationStore.telemetryService,