+++ packages/legend-application-pure-ide/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @finos/legend-application-pure-ide
 
+## 6.2.43
+
+## 6.2.42
+
+## 6.2.41
+
 ## 6.2.40
 
 ## 6.2.39