+++ packages/legend-application-query/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @finos/legend-application-query
 
+## 10.0.3
+
+## 10.0.2
+
+## 10.0.1
+
 ## 10.0.0
 
 ### Major Changes