+++ packages/legend-application-taxonomy/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @finos/legend-application-taxonomy
 
+## 6.2.46
+
+## 6.2.45
+
+## 6.2.44
+
 ## 6.2.43
 
 ## 6.2.42