+++ packages/legend-application/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @finos/legend-application
 
+## 11.0.3
+
+## 11.0.2
+
+## 11.0.1
+
 ## 11.0.0
 
 ### Major Changes