+++ packages/legend-extension-dsl-diagram/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @finos/legend-extension-dsl-diagram
 
+## 7.1.37
+
+## 7.1.36
+
+## 7.1.35
+
 ## 7.1.34
 
 ## 7.1.33