+++ packages/legend-extension-dsl-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @finos/legend-extension-dsl-service
 
+## 0.1.33
+
+## 0.1.32
+
+## 0.1.31
+
 ## 0.1.30
 
 ### Patch Changes