+++ packages/legend-extension-format-json-schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @finos/legend-extension-format-json-schema
 
+## 1.0.38
+
+## 1.0.37
+
+## 1.0.36
+
 ## 1.0.35
 
 ## 1.0.34