+++ packages/legend-extension-format-morphir/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @finos/legend-extension-format-morphir
 
+## 1.0.50
+
+## 1.0.49
+
+## 1.0.48
+
 ## 1.0.47
 
 ## 1.0.46