+++ packages/legend-extension-store-flat-data/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @finos/legend-extension-store-flat-data
 
+## 0.0.52
+
+## 0.0.51
+
+## 0.0.50
+
 ## 0.0.49
 
 ## 0.0.48