+++ packages/legend-extension-tracer-zipkin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @finos/legend-extension-tracer-zipkin
 
+## 3.0.39
+
+## 3.0.38
+
+## 3.0.37
+
 ## 3.0.36
 
 ## 3.0.35