+++ packages/legend-graph/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @finos/legend-graph
 
+## 23.0.3
+
+### Patch Changes
+
+- [#1963](https://github.com/finos/legend-studio/pull/1963) [`1fbccf052`](https://github.com/finos/legend-studio/commit/1fbccf052b7e2048f54e089ffb3f7848e434a9fe) ([@gayathrir11](https://github.com/gayathrir11)) - Fix calling `compile` api twice for compilation.
+
+## 23.0.2
+
+## 23.0.1
+
+### Patch Changes
+
+- [#1953](https://github.com/finos/legend-studio/pull/1953) [`1a96a29fd`](https://github.com/finos/legend-studio/commit/1a96a29fdf60db8757e85b23f4554b2729e1606b) ([@YannanGao-gs](https://github.com/YannanGao-gs)) - Fix a bug in which the generation of assertion data produces a misleading error message.
+
 ## 23.0.0
 
 ### Major Changes