+++ packages/legend-shared/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @finos/legend-shared
 
+## 7.0.0
+
+### Major Changes
+
+- [#1973](https://github.com/finos/legend-studio/pull/1973) [`830934048`](https://github.com/finos/legend-studio/commit/8309340482c7a1aa21fedbd90063cd0c521a5ddf) ([@akphi](https://github.com/akphi)) - **BREAKING CHANGE:** Moved `EventNotifierService` and `TelemetryService` to `@finos/legend-application` and reshaped `TelemetryService` as well as `TelemetryServicePlugin` to take a `setup()` method for instead of just allowing setting the user ID via `setUserId()`.
+
+### Patch Changes
+
+- [#1973](https://github.com/finos/legend-studio/pull/1973) [`830934048`](https://github.com/finos/legend-studio/commit/8309340482c7a1aa21fedbd90063cd0c521a5ddf) ([@akphi](https://github.com/akphi)) - Fix a bug with `StopWatch` elapsed is not properly calculated.
+
+## 6.2.23
+
+### Patch Changes
+
+- [#1960](https://github.com/finos/legend-studio/pull/1960) [`52c944f93`](https://github.com/finos/legend-studio/commit/52c944f93815d790263711c261ca5514d74b6b28) ([@xannem](https://github.com/xannem)) - Fix pretty const name.
+
+## 6.2.22
+
 ## 6.2.21
 
 ## 6.2.20