+++ packages/stylelint-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/stylelint-config-legend-studio
 
+## 2.0.31
+
+## 2.0.30
+
 ## 2.0.29
 
 ## 2.0.28